// src/features/common/activity-log/enums/activity-action-type.enum.ts
export enum ActivityActionType {
  SUBMIT_APPLICATION = 'SUBMIT_APPLICATION',
  APPROVE_APPLICATION = 'APPROVE_APPLICATION',
  REJECT_APPLICATION = 'REJECT_APPLICATION',
  PAY_ANNUAL_FEE = 'PAY_ANNUAL_FEE',
  MEMBER_ACTIVATED = 'MEMBER_ACTIVATED',
  MEMBER_SUSPENDED = 'MEMBER_SUSPENDED',
  PAY_MULTI_YEARS = 'PAY_MULTI_YEARS',
  EDIT_BASIC_INFO = 'EDIT_BASIC_INFO',
  UPDATE_ATTACHMENT = 'UPDATE_ATTACHMENT',
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
  CHANGE_PASSWORD = 'CHANGE_PASSWORD',
  RESET_PASSWORD = 'RESET_PASSWORD',
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  DELETE_USER = 'DELETE_USER',
  UPLOAD_FILE = 'UPLOAD_FILE',
  DELETE_FILE = 'DELETE_FILE',
}